refactor [structure] :seperate app from main for integration test ease
