+++ src/main.js
@@ -1,32 +1,16 @@
-import express from "express";
+import dotenv from "dotenv";
 
-import bodyParser from "body-parser";
-
-import { tenantRouter, vehicleRouter } from "./routes/index.js";
-
+import app from "./app.js";
 import { db } from "./repository/index.js";
 
-const app = express();
-app.use(bodyParser.json());
+// Load environment variables from .env file
+dotenv.config();
 
 // Initialize the database connection
 db.init();
 
-// Express API routes for CRUD operations
-app.use("/api/v1/", [tenantRouter, vehicleRouter]);
-
-// catch all errors
-app.use((error, req, res, next) => {
-  //   console.error("here", error.statusCode, error.message, error.stack);
-  res
-    .status(error.statusCode || 500)
-    .json({ message: error.message ?? "Internal Server Error" });
-  next();
-});
-
-export default app;
-
 // Start the Express server
+
 const port = process.env.PORT || 3000;
 app.listen(port, () => {
   console.log(`Server is running on port ${port}`);
